Revert "feat: lemon file input (#12830)"
This reverts commit 0da5ba426f26f31ddd93b9cffd1cd0561e3b182d.
