--- frontend/src/lib/components/LemonFileInput/LemonFileInput.scss
@@ -1,7 +1,0 @@
-.FileDropTarget {
-    border: 3px dashed transparent;
-}
-
-.FileDropTarget--active {
-    border-color: var(--primary);
-}
