+++ frontend/src/styles/utilities.scss
@@ -864,7 +864,3 @@
 .z-auto {
     z-index: auto;
 }
-
-.hidden {
-    display: none;
-}
